Extracted the reload amd and the mob races from DB.
