+++ src/cmd/score.cpp
@@ -12,7 +12,7 @@
 #include "communication/parcel.h"
 #include "entities/char_data.h"
 #include "entities/char_player.h"
-#include "entities/player_races.h"
+#include "game_mechanics/player_races.h"
 #include "game_fight/fight_hit.h"
 #include "game_mechanics/bonus.h"
 #include "game_mechanics/glory.h"
