+++ src/heartbeat.cpp
@@ -410,9 +410,9 @@ Heartbeat::steps_t &pulse_steps() {
 							 40,
 							 std::make_shared<SimpleCall>(Clan::ClanSave)),
 		Heartbeat::PulseStep("Celebrates: sanitize",
-							 Celebrates::CLEAN_PERIOD * 60 * kPassesPerSec,
+							 celebrates::CLEAN_PERIOD * 60 * kPassesPerSec,
 							 39,
-							 std::make_shared<SimpleCall>(Celebrates::sanitize)),
+							 std::make_shared<SimpleCall>(celebrates::sanitize)),
 		Heartbeat::PulseStep("Record usage", 5 * 60 * kPassesPerSec, 37, std::make_shared<SimpleCall>(record_usage)),
 		Heartbeat::PulseStep("Reload proxy ban",
 							 5 * 60 * kPassesPerSec,
