+++ src/structs/global_objects.h
@@ -69,13 +69,13 @@ class GlobalObjects {
 	static OutputThread &output_thread();
 	static ZoneTable &zone_table();
 
-	static Celebrates::CelebrateList &mono_celebrates();
-	static Celebrates::CelebrateList &poly_celebrates();
-	static Celebrates::CelebrateList &real_celebrates();
-	static Celebrates::CelebrateMobs &attached_mobs();
-	static Celebrates::CelebrateMobs &loaded_mobs();
-	static Celebrates::CelebrateObjs &attached_objs();
-	static Celebrates::CelebrateObjs &loaded_objs();
+	static celebrates::CelebrateList &mono_celebrates();
+	static celebrates::CelebrateList &poly_celebrates();
+	static celebrates::CelebrateList &real_celebrates();
+	static celebrates::CelebrateMobs &attached_mobs();
+	static celebrates::CelebrateMobs &loaded_mobs();
+	static celebrates::CelebrateObjs &attached_objs();
+	static celebrates::CelebrateObjs &loaded_objs();
 
 	static GlobalTriggersStorage &trigger_list();
 	static BloodyInfoMap &bloody_map();
